requirements.txt
tinytuya>=1.13.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
